requirements.txt
absl-py==1.4.0
ackermann-drive-teleop==1.0.1
action-tutorials-py==0.14.4
altair==4.2.2
ament-cmake-test==1.1.6
ament-copyright==0.10.7
ament-cppcheck==0.10.7
ament-cpplint==0.10.7
ament-flake8==0.10.7
ament-index-python==1.2.0
ament-lint==0.10.7
ament-lint-cmake==0.10.7
ament-package==0.13.1
ament-pep257==0.10.7
ament-uncrustify==0.10.7
ament-xmllint==0.10.7
angles==1.14.0
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
blinker==1.5
cachetools==5.3.0
car-keyop==2.0.0
catboost==1.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cryptography==39.0.1
cv-bridge==3.1.3
cycler==0.11.0
decorator==5.1.1
demo-nodes-py==0.14.4
domain-coordinator==0.9.2
entrypoints==0.4
examples-rclpy-executors==0.11.2
examples-rclpy-minimal-action-client==0.11.2
examples-rclpy-minimal-action-server==0.11.2
examples-rclpy-minimal-client==0.11.2
examples-rclpy-minimal-publisher==0.11.2
examples-rclpy-minimal-service==0.11.2
examples-rclpy-minimal-subscriber==0.11.2
flatbuffers==23.1.21
fonttools==4.38.0
frozendict==2.3.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.8.0
html5lib==1.1
id-track-utils==0.6.0
idna==3.4
image-geometry==3.1.3
imperial-driverless-utils==1.0.3
importlib-metadata==6.0.0
importlib-resources==5.10.2
interactive-markers==2.2.0
Jinja2==3.1.2
joblib==1.2.0
joint-state-publisher==2.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
launch==0.17.2
launch-ros==0.14.5
launch-testing==0.17.2
launch-testing-ros==0.14.5
launch-xml==0.17.2
launch-yaml==0.17.2
libclang==15.0.6.1
lidar-only-cone-detector==0.1.2
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
multitasking==0.0.11
nav2-common==1.0.12
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
osrf-pycommon==2.0.0
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
patsy==0.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plotly==5.13.0
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-qt-binding==1.0.8
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
qt-dotgraph==2.0.1
qt-gui==2.0.1
qt-gui-cpp==2.0.1
qt-gui-py-common==2.0.1
quality-of-service-demo-py==0.14.4
rclpy==1.9.3
rcutils==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
resource-retriever==2.5.0
rich==13.3.1
ros2action==0.13.5
ros2bag==0.9.2
ros2cli==0.13.5
ros2component==0.13.5
ros2doctor==0.13.5
ros2interface==0.13.5
ros2launch==0.14.5
ros2lifecycle==0.13.5
ros2multicast==0.13.5
ros2node==0.13.5
ros2param==0.13.5
ros2pkg==0.13.5
ros2run==0.13.5
ros2service==0.13.5
ros2topic==0.13.5
rosbag2-py==0.9.2
rosidl-adapter==2.2.3
rosidl-cli==2.2.3
rosidl-cmake==2.2.3
rosidl-generator-c==2.2.3
rosidl-generator-cpp==2.2.3
rosidl-generator-py==0.11.4
rosidl-parser==2.2.3
rosidl-runtime-py==0.9.1
rosidl-typesupport-c==1.2.1
rosidl-typesupport-cpp==1.2.1
rosidl-typesupport-fastrtps-c==1.2.1
rosidl-typesupport-fastrtps-cpp==1.2.1
rosidl-typesupport-introspection-c==2.2.3
rosidl-typesupport-introspection-cpp==2.2.3
rpyutils==0.2.0
rqt-action==2.0.1
rqt-bag==1.1.2
rqt-bag-plugins==1.1.2
rqt-console==2.0.2
rqt-graph==1.2.2
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.1.0
rqt-plot==1.1.1
rqt-publisher==1.1.3
rqt-py-common==1.1.2
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.4.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
semver==2.13.0
six==1.16.0
smclib==3.0.1
smmap==5.0.0
soupsieve==2.3.2.post1
sros2==0.10.3
statsmodels==0.13.5
streamlit==1.18.1
streamlit-option-menu==0.3.2
teleop-twist-keyboard==2.3.2
tenacity==8.2.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tf2-geometry-msgs==0.17.5
tf2-kdl==0.17.5
tf2-py==0.17.5
tf2-ros-py==0.17.5
tf2-tools==0.17.5
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
topic-monitor==0.14.4
topic-tools==1.0.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xacro==2.0.7
xgboost==1.7.3
yfinance==0.2.11
zipp==3.13.0
